+++ packages/explorer-consumer/src/consumers/ExplorerConsumer.ts
@@ -1,0 +1,5 @@
+import BaseConsumer from '../BaseConsumer'
+
+export class ExplorerConsumer extends BaseConsumer {
+  //
+}