+++ packages/networks/src/endpoints.ts
@@ -10,39 +10,39 @@ export const urlEndpointsMainnet: UrlEndpoint = {
 }
 
 export const urlEndpointsPublic: UrlEndpoint = {
-  baseUrl: 'https://public-api.injective.network/api',
-  chainUrl: 'https://public-grpc.injective.network',
-  chainHttpUrl: 'https://public-lcd.injective.network',
-  exchangeUrl: 'https://public-api.injective.network',
-  exchangeGatewayUrl: 'https://public-gateway.injective.network',
-  explorerUrl: 'https://public-api.injective.network/api/explorer/v1',
+  baseUrl: 'https://public.api.injective.network/api',
+  chainUrl: 'https://public.grpc.injective.network',
+  chainHttpUrl: 'https://public.lcd.injective.network',
+  exchangeUrl: 'https://public.api.injective.network',
+  exchangeGatewayUrl: 'https://public.web3-gateway.injective.network',
+  explorerUrl: 'https://public.api.injective.network/api/explorer/v1',
 }
 
 export const urlEndpointsMainnetOld: UrlEndpoint = {
-  baseUrl: 'https://staking-api.injective.network/api',
-  chainUrl: 'https://staking-grpc.injective.network',
-  chainHttpUrl: 'https://staking-lcd.injective.network',
-  exchangeUrl: 'https://staking-api.injective.network',
-  exchangeGatewayUrl: 'https://staking-gateway.injective.network',
-  explorerUrl: 'https://staking-api.injective.network/api/explorer/v1',
+  baseUrl: 'https://api.injective.network/api',
+  chainUrl: 'https://grpc.injective.network',
+  chainHttpUrl: 'https://lcd.injective.network',
+  exchangeUrl: 'https://api.injective.network',
+  exchangeGatewayUrl: 'https://web3-gateway.injective.network',
+  explorerUrl: 'https://api.injective.network/api/explorer/v1',
 }
 
 export const urlEndpointsStaking: UrlEndpoint = {
-  baseUrl: 'https://staking-api-testnet.injective.network/api',
-  chainUrl: 'https://staking-grpc-testnet.injective.network',
-  chainHttpUrl: 'https://staking-lcd-testnet.injective.network',
-  exchangeUrl: 'https://staking-api-testnet.injective.network',
-  exchangeGatewayUrl: 'https://staking-gateway-testnet.injective.network',
-  explorerUrl: 'https://staking-api-testnet.injective.network/api/explorer/v1',
+  baseUrl: 'https://testnet.api.injective.dev/api',
+  chainUrl: 'https://testnet.grpc.injective.dev',
+  chainHttpUrl: 'https://testnet.lcd.injective.dev',
+  exchangeUrl: 'https://testnet.api.injective.dev',
+  exchangeGatewayUrl: 'https://testnet.web3-gateway.injective.dev',
+  explorerUrl: 'https://testnet.api.injective.dev/api/explorer/v1',
 }
 
 export const urlEndpointsDevnet: UrlEndpoint = {
-  baseUrl: 'https://devnet-api.injective.dev/api',
-  chainUrl: 'https://devnet-grpc.injective.dev',
-  chainHttpUrl: 'https://devnet-lcd.injective.dev',
-  exchangeUrl: 'https://devnet-api.injective.dev',
-  exchangeGatewayUrl: 'https://devnet-gateway.injective.dev',
-  explorerUrl: 'https://devnet-api.injective.dev/api/explorer/v1',
+  baseUrl: 'https://devnet.api.injective.dev/api',
+  chainUrl: 'https://devnet.grpc.injective.dev',
+  chainHttpUrl: 'https://devnet.lcd.injective.dev',
+  exchangeUrl: 'https://devnet.api.injective.dev',
+  exchangeGatewayUrl: 'https://devnet.gateway.injective.dev',
+  explorerUrl: 'https://devnet.api.injective.dev/api/explorer/v1',
 }
 
 export const urlEndpointsLocal: UrlEndpoint = {