+++ packages/derivatives-consumer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.48](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/derivatives-consumer@0.0.47...@injectivelabs/derivatives-consumer@0.0.48) (2021-09-16)
+
+
+### Features
+
+* explorer consumer package ([1499017](https://github.com/InjectiveLabs/injective-ts/commit/1499017bca3d09f74e5a0eecedeba8d445d37215))
+
+
+
+
+
 ## [0.0.47](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/derivatives-consumer@0.0.46...@injectivelabs/derivatives-consumer@0.0.47) (2021-09-13)