+++ packages/exchange-consumer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.88](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exchange-consumer@0.0.87...@injectivelabs/exchange-consumer@0.0.88) (2021-09-16)
+
+
+### Features
+
+* explorer consumer package ([1499017](https://github.com/InjectiveLabs/injective-ts/commit/1499017bca3d09f74e5a0eecedeba8d445d37215))
+
+
+
+
+
 ## [0.0.87](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exchange-consumer@0.0.86...@injectivelabs/exchange-consumer@0.0.87) (2021-09-13)